POI：bug 59748: replace Hashtable with HashMap; contributed by Axel Howind

+++ ooxml/XSSF/UserModel/XSSFRow.cs
@@ -85,7 +85,7 @@ namespace NPOI.XSSF.UserModel
                 {
                     nextRowNum = 1;
                 }
-                row.r = (nextRowNum);
+                row.r = (uint)nextRowNum;
             }
         }
